requirements.txt
python-dateutil
requests
selectorlib
beautifulsoup4